add stock prices endpoint

+++ stock-service/src/main/java/seng468scalability/com/stock_transactions/request/InternalDeleteStockTXRequest.java
@@ -2,4 +2,4 @@ package seng468scalability.com.stock_transactions.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record InternalDeleteStockTXRequest(@JsonProperty("stock_tx_id") Integer stockTXId) {}
+public record InternalDeleteStockTXRequest(@JsonProperty("stock_tx_id") Long stockTXId) {}
